+++ src/components/calcite-video/calcite-video.e2e.ts
@@ -1,0 +1,18 @@
+import { newE2EPage } from "@stencil/core/testing";
+import { accessible, renders } from "../../tests/commonTests";
+
+describe("calcite-chip", () => {
+  //  it("renders", async () => renders("<calcite-video"));
+  /*
+  it("is accessible", async () =>
+    accessible(`<calcite-video>`));
+    */
+});
+
+// scrubbing works (check updated time val)
+// pause / play works
+// fullscreen works enter and exit
+// play on mouseenter, pause on mouseleave
+// pause on scrubber pauses video
+// playing another video pauses playing video
+// loading shows until ready state is good