+++ stencil.config.ts
@@ -26,6 +26,7 @@ export const config: Config = {
         "calcite-dropdown-item",
       ],
     },
+    { components: ["calcite-graph"] },
     { components: ["calcite-icon"] },
     { components: ["calcite-input"] },
     { components: ["calcite-input-message"] },
@@ -37,9 +38,11 @@ export const config: Config = {
     { components: ["calcite-pagination"] },
     { components: ["calcite-popover", "calcite-popover-manager"] },
     { components: ["calcite-progress"] },
+    { components: ["calcite-radio-button", "calcite-radio-button-group"] },
     { components: ["calcite-radio-group", "calcite-radio-group-item"] },
     { components: ["calcite-scrim"] },
     { components: ["calcite-slider"] },
+    { components: ["calcite-split-button"] },
     { components: ["calcite-stepper", "calcite-stepper-item"] },
     { components: ["calcite-switch"] },
     {
@@ -52,6 +55,7 @@ export const config: Config = {
     },
     { components: ["calcite-tooltip", "calcite-tooltip-manager"] },
     { components: ["calcite-tree", "calcite-tree-item"] },
+    { components: ["calcite-video"] },
   ],
   outputTargets: [
     { type: "dist-hydrate-script" },